Deprecate internal libraries
Deprecate data internal library. MedHash data internal library is to be
deprecated at the release of v0.4.0. data will be removed with the
release of v0.6.0.

Deprecate packageinfo internal library. MedHash packageinfo internal
library is to be deprecated at the release of v0.4.0. packageinfo will
be removed with the release of v0.6.0.