Manifest Specification v0.4.0
Italicize Manifest. In the context of this Specification, Manifest is a
proper noun.

Italicize Media. In the context of this Specification, Media is a proper
noun.

Italicize Hash Container. In the context of this Specification, Hash
Container is a proper noun.

Remove version numbering scheme guideline. This guideline concerns the
Specification maintainers, not implementers. Additionally, this
guideline concerns future versions of the Specification. Current version
of the Specification should influence, not dictate, future versions of
the Specification.

Add generator field. This is an optional field for generators to declare
and credit themselves on the Manifest.

Add guideline on path separator. For simplicity, path separators for the
Manifest should be unified. Translation between the Specification path
separator and the host OS path separator is an implementation detail.

Add support for SHA3-256. SHA3-256 is more modern and more secure than
SHA256. For the most part, this algorithm is intended to be a drop-in
replacement of SHA256.

Deprecate SHA1 and MD5. These algorithm were included for compatibility
reasons. They should remain supported by implementers, but they should
be deprecated. SHA256 and SHA3-256 should be preferred over SHA1 and
MD5.