Rewrite medhash in v2
Package medhash/v2 rewrites the MedHash library. Hash generation and
verification are done at the library level, rather than needing to be
set up by the library consumer. Additionally, hashing is done on media
addition.

Note: this rewrite is versioned under v2 for now, but we will eventually
replace the root medhash package in a backwards incompatible way. This
will be done prior to merging to main.