Update main to work with cmd/v2
Package medhash/v2 rewrites the MedHash library. Package cmd/v2 is an
updated commands package that works with this new version of the
library. Update main to work with cmd/v2.

Note: this update is versioned under v2 for now, but we will eventually
replace the root main in a backwards incompatible way. This will be done
prior to merging to main.